modified ricette handling and added snackbar on save profile changes

+++ app/src/main/java/com/example/jet_bring/ui/ispirazione/addRicettaViewModel.kt
@@ -63,7 +63,7 @@ class AddRicettaViewModel : ViewModel() {
 
     fun onSaveDone() {
 
-        ricetta = Ricetta(id = (ricette.lastIndex+ 1), titolo = ricetta.titolo, descrizione = ricetta.descrizione, pubblicatore = ricetta.pubblicatore, immagine = R.drawable.empty_plate, sourceUrl = ricetta.sourceUrl, voti = ricetta.voti, ingredienti = listeViewModel.selectedProducts.component1()
+        ricetta = Ricetta(id = (ricette.lastIndex.toLong()+1), titolo = ricetta.titolo, descrizione = ricetta.descrizione, pubblicatore = ricetta.pubblicatore, immagine = R.drawable.empty_plate, sourceUrl = ricetta.sourceUrl, voti = ricetta.voti, ingredienti = listeViewModel.selectedProducts.component1()
 
         )
